refactor: remove button type from core content and extend content for telegram with button support
